Move new route files up into the routes/ folder.
